Creature controller unification starts
